기존 batteryQI 네임스페이스로 저장되어있던것을 batteryQI_plus로 변경. Model에서 네임스페이스 Rename 변경하니까 문제없이 변경 완료. + Setting View 텍스트 박스 위치 수정
