+++ batteryQI_plus/App.xaml
@@ -1,9 +1,9 @@
-﻿<Application x:Class="batteryQI.App"
+﻿<Application x:Class="batteryQI_plus.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:fa="http://schemas.awesome.incremented/wpf/xaml/fontawesome.sharp"
-             xmlns:local="clr-namespace:batteryQI"
-             xmlns:locator="clr-namespace:batteryQI.ViewModels"
+             xmlns:local="clr-namespace:batteryQI_plus"
+             xmlns:locator="clr-namespace:batteryQI_plus.ViewModels"
              >
 
     <Application.Resources>
