+++ batteryQI_plus/ViewModels/Bases/ViewModelBases.cs
@@ -6,12 +6,12 @@ using System.Threading.Tasks;
 using System.Windows;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
-using batteryQI.Models;
-using batteryQI.Views;
+using batteryQI_plus.Models;
+using batteryQI_plus.Views;
 using System.Windows.Controls;
 using System.Data.Common;
 
-namespace batteryQI.ViewModels.Bases
+namespace batteryQI_plus.ViewModels.Bases
 {
     public partial class ViewModelBases : ObservableObject, IDisposable
     {
