+++ batteryQI_plus/ViewModels/InspectViewModelIlist.cs
@@ -3,11 +3,11 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using batteryQI.Models;
+using batteryQI_plus.Models;
 using System.Windows;
-using batteryQI.ViewModels.Bases;
+using batteryQI_plus.ViewModels.Bases;
 
-namespace batteryQI.ViewModels
+namespace batteryQI_plus.ViewModels
 {
     public partial class InspectViewModel : ViewModelBases
     {
