+++ batteryQI_plus/Views/CustomMessageBoxYesNo.xaml
@@ -1,9 +1,9 @@
-﻿<Window x:Class="batteryQI.Views.CustomMessageBoxYesNo"
+﻿<Window x:Class="batteryQI_plus.Views.CustomMessageBoxYesNo"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:batteryQI.Views"
+        xmlns:local="clr-namespace:batteryQI_plus.Views"
         WindowStyle="None" ResizeMode="NoResize"
         Height="250" Width="400"
         SizeToContent="WidthAndHeight" Background="White"
