+++ batteryQI_plus/Views/ErrorInfoView.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="batteryQI.Views.ErrorInfoView"
+﻿<Window x:Class="batteryQI_plus.Views.ErrorInfoView"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         Height="600" Width="880"
