+++ batteryQI_plus/Views/InspectionImage.xaml.cs
@@ -11,10 +11,10 @@ using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Shapes;
-using batteryQI.ViewModels;
-using batteryQI.Views.UserControls;
+using batteryQI_plus.ViewModels;
+using batteryQI_plus.Views.UserControls;
 
-namespace batteryQI.Views
+namespace batteryQI_plus.Views
 {
     /// <summary>
     /// Interaction logic for InspectionImage.xaml
