+++ batteryQI_plus/Views/LoginWindow.xaml
@@ -1,8 +1,8 @@
-﻿<Window x:Class="batteryQI.Views.LoginWindow"
+﻿<Window x:Class="batteryQI_plus.Views.LoginWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:local="clr-namespace:batteryQI.Views"
-        xmlns:vm="clr-namespace:batteryQI.ViewModels.Bases"
+        xmlns:local="clr-namespace:batteryQI_plus.Views"
+        xmlns:vm="clr-namespace:batteryQI_plus.ViewModels.Bases"
         Title="LoginWindow" Height="450" Width="800"
         >
     <Window.DataContext>
