+++ batteryQI_plus/Views/MainWindow.xaml
@@ -2,8 +2,8 @@
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:fa="http://schemas.awesome.incremented/wpf/xaml/fontawesome.sharp"
-        xmlns:viewmodels="clr-namespace:batteryQI.ViewModels"
-        xmlns:Default="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="Default" x:Class="batteryQI.Views.MainWindow"
+        xmlns:viewmodels="clr-namespace:batteryQI_plus.ViewModels"
+        xmlns:Default="http://schemas.microsoft.com/expression/blend/2008" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" mc:Ignorable="Default" x:Class="batteryQI_plus.Views.MainWindow"
         Title="BatteryQI"         
         WindowStartupLocation="CenterScreen" 
         WindowState="Maximized"
