+++ batteryQI_plus/Views/UserControls/AnalysisView.xaml
@@ -1,8 +1,8 @@
-﻿<UserControl x:Class="batteryQI.Views.UserControls.AnalysisView"
+﻿<UserControl x:Class="batteryQI_plus.Views.UserControls.AnalysisView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:batteryQI.Extensions"
-             xmlns:vm="clr-namespace:batteryQI.ViewModels"
+             xmlns:local="clr-namespace:batteryQI_plus.Extensions"
+             xmlns:vm="clr-namespace:batteryQI_plus.ViewModels"
              >
     <UserControl.DataContext>
         <Binding Path="TabControlViewModel" Source="{StaticResource Locator}" />
