+++ batteryQI_plus/Views/UserControls/DashboardView.xaml
@@ -1,7 +1,7 @@
-﻿<UserControl x:Class="batteryQI.Views.UserControls.DashboardView"
+﻿<UserControl x:Class="batteryQI_plus.Views.UserControls.DashboardView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:batteryQI.ViewModels"
+             xmlns:local="clr-namespace:batteryQI_plus.ViewModels"
              >
     <!-- DataContext 설정 -->
     <UserControl.DataContext>
