+++ batteryQI_plus/Views/UserControls/ErrorInspection.xaml.cs
@@ -12,9 +12,9 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
-using batteryQI.ViewModels;
+using batteryQI_plus.ViewModels;
 
-namespace batteryQI.Views.UserControls
+namespace batteryQI_plus.Views.UserControls
 {
     /// <summary>
     /// Interaction logic for ErrorInspection.xaml
