+++ batteryQI_plus/Views/UserControls/ErrorReason.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="batteryQI.Views.UserControls.ErrorReason"
+﻿<UserControl x:Class="batteryQI_plus.Views.UserControls.ErrorReason"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml">
