+++ batteryQI_plus/Views/UserControls/ErrorReason.xaml.cs
@@ -12,10 +12,10 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
-using batteryQI.Views;
-using batteryQI.ViewModels;
+using batteryQI_plus.Views;
+using batteryQI_plus.ViewModels;
 
-namespace batteryQI.Views.UserControls
+namespace batteryQI_plus.Views.UserControls
 {
     /// <summary>
     /// Interaction logic for ErrorReason.xaml
