+++ batteryQI_plus/Views/UserControls/SettingView.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="batteryQI.Views.UserControls.SettingView"
+﻿<UserControl x:Class="batteryQI_plus.Views.UserControls.SettingView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -80,7 +80,7 @@
                                             <Border BorderBrush="Gray" BorderThickness="0,0,0,5" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"/>
                                             <ToggleButton Style="{StaticResource OnOffToggleButton}" Margin="0,20,0,30" Width="170" Height="50" HorizontalAlignment="Left"
                                                               IsChecked="{Binding IsLinePower, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"/>
-                                            <TextBlock Text="생산라인 전원" FontSize="22" FontWeight="SemiBold" HorizontalAlignment="Right" VerticalAlignment="Center" Margin="0,20,760,30"/>
+                                            <TextBlock Text="생산라인 전원" FontSize="22" FontWeight="SemiBold" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="190,20,0,30"/>
 
                                         </Grid>
 
@@ -111,7 +111,7 @@
                                             <Border BorderBrush="Gray" BorderThickness="0,0,0,5" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"/>
                                             <ToggleButton Style="{StaticResource OnOffToggleButton}" Margin="0,20,0,30" Width="170" Height="50" HorizontalAlignment="Left"
                                                               IsChecked="{Binding IsLinePower, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}"/>
-                                            <TextBlock Text="생산라인 전원" FontSize="22" FontWeight="SemiBold" HorizontalAlignment="Right" VerticalAlignment="Center" Margin="0,20,760,30"/>
+                                            <TextBlock Text="생산라인 전원" FontSize="22" FontWeight="SemiBold" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="190,20,0,30"/>
 
                                         </Grid>
