+++ batteryQI_plus/Views/UserControls/SettingView.xaml.cs
@@ -13,9 +13,9 @@ using System.Windows.Media;
 using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
-using batteryQI.ViewModels;
+using batteryQI_plus.ViewModels;
 
-namespace batteryQI.Views.UserControls
+namespace batteryQI_plus.Views.UserControls
 {
     /// <summary>
     /// Interaction logic for ManagerPage.xaml
